Refactor PromoCard UI + extend FluffyCard
