+++ src/components/PromoCard/PromoCard.md
@@ -1,15 +1,11 @@
 # PromoCard
 
-A PromoCard component is used to encapsulate pieces of UI that share a common concept. It is an enhanced version of [Card](../Card) that includes a colored top border.
+A PromoCard component is used to encapsulate pieces of UI that share a common concept. It is an enhanced version of [FluffyCard](../FluffyCard) that includes a colored top border.
 
 ## Example
 
 ```jsx
-<PromoCard promoColor="66CDAA">
-  Lorem ipsum dolor sit amet, consectetur adipisicing elit. Deserunt iste soluta
-  consequuntur, obcaecati labore nam cumque repellendus similique aliquam ea
-  architecto nobis aliquid beatae rerum quos fugit expedita asperiores quaerat!
-</PromoCard>
+<PromoCard promoColor="red">Mugatu!</PromoCard>
 ```
 
 ## Props
@@ -17,6 +13,6 @@ A PromoCard component is used to encapsulate pieces of UI that share a common co
 | Prop       | Type     | Default | Description                     |
 | ---------- | -------- | ------- | ------------------------------- |
 | className  | `string` |         | The className of the component. |
-| promoColor | `string` | blue    | The displayed top border color. |
+| promoColor | `string` | `blue`  | The displayed top border color. |
 
-For additional props, check out [Card](../Card).
+For additional props, check out [FluffyCard](../FluffyCard).
